Fix unused warnings
The intent of "scalacOption in compile .." is to enable "Ywarn-unused"
and "Ywarn-unused-import" when compiling, but not when in the REPL.

However in sbt, because of how setting delegation works (and/or how
things are actually defined by default), task scope is a second-class
citizen to configuration scope, so "in compile" is not enough, you need
"in Compile in compile" (in this case via "inBoth").

Also add a few more scalac options on advice by scalac, without scoping.